feat(actions): warn about references when deleting from actions list
Extract deleteActionWithWarning utility that fetches references and
shows a confirmation dialog before deleting. Used by both the actions
list table and the action detail page.
